Partway through implementation - BROKEN BUILD
The state of the code for this commit is broken. The reason I'm
committing it is that real life stuff (my wife being ill) is going
on at the moment, and I'm not going to get a chance to work on
this for a few days. I wanted to get a version of where I'm at
up on Github just in case.